switcharoo - only require serdes when needed over storing until needed.
the streams were holding on to the correct serdes until they were
needed, eg join, left_join, windowed, materialized.
Because of this we had to set the correct serde on every map and rekey,
and pass it between different stream states. When we flipped this, the
code base became simpler